+++ src/components/analytics/YandexMetrika.tsx
@@ -7,7 +7,7 @@ export function YandexMetrika() {
   return (
     <Script
       id="yandex-metrika"
-      strategy="afterInteractive"
+      strategy="lazyOnload"
     >{`
       (function(m,e,t,r,i,k,a){m[i]=m[i]||function(){(m[i].a=m[i].a||[]).push(arguments)};
       m[i].l=1*new Date();